+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Changelog by https://keepachangelog.com
 - EditorConfig support.
+- Uptime value in ping check view.
 
 ### Changed
 - Gunicorn daemon runs with UVloop worker class.